requirements.txt
playwright
boto3
sqlalchemy
psycopg2-binary
python-dotenv
